requirements.txt
Flask==2.3.3
python-dotenv
google-cloud-storage
pymysql
gunicorn